feat: add player pxs system

+++ contracts/src/models/proposal.cairo
@@ -50,7 +50,7 @@ struct PlayerVote {
 struct PixelRecoveryRate {
     #[key]
     id: usize,
-    rate: u32
+    rate: u64
 }
 
 impl ProposalTypeFelt252 of Into<ProposalType, felt252> {
